refactor: Curation 컨트롤러,서비스 분리 및 테스트 수정
