+++ src/main/java/net/pointofviews/curation/controller/CurationMemberController.java
@@ -1,0 +1,11 @@
+package net.pointofviews.curation.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/api/movies/curations")
+@RequiredArgsConstructor
+public class CurationMemberController implements CurationMemberSpecification {
+
+}
