+++ src/main/java/net/pointofviews/curation/controller/CurationMemberSpecification.java
@@ -1,0 +1,8 @@
+package net.pointofviews.curation.controller;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+
+@Tag(name = "Curation", description = "사용자 영화 추천 큐레이션 API")
+public interface CurationMemberSpecification {
+
+}
