+++ src/main/java/net/pointofviews/curation/service/CurationService.java
@@ -6,10 +6,6 @@ import net.pointofviews.curation.dto.response.ReadCurationListResponse;
 import net.pointofviews.curation.dto.response.ReadCurationResponse;
 
 public interface CurationService {
-    ReadCurationResponse saveCuration(CreateCurationRequest request);
     ReadCurationListResponse readAllCurations();
     ReadCurationResponse readCuration(Long curationId);
-    ReadCurationListResponse searchCurations(String theme, CurationCategory category);
-    ReadCurationResponse updateCuration(Long curationId, CreateCurationRequest request);
-    void deleteCuration(Long curationId);
 }
